Add cell balancing

+++ firmware/src/controller/power.cpp
@@ -11,7 +11,7 @@
 #include <Arduino.h>
 
 
-uint16_t update_charging_status(charging_status_t *charging_status, const Adafruit_INA260 *ina260) {
+uint8_t update_charging_status(charging_status_t *charging_status, const Adafruit_INA260 *ina260) {
 
     // Make a copy before updating for comparison
     charging_status_t old_charging_status = *charging_status;
